complete

+++ src/app/modules/admin/admin.service.ts
@@ -1,7 +1,15 @@
 import { UserRegister } from '../auth/auth.model';
+import { currentUserEmail } from '../auth/auth.utils';
 import { BlogPost } from '../blogs/blog.model';
 
-const blockUserByAdminIntoDb = async (userId: string) => {
+const blockUserByAdminIntoDB = async (userId: string) => {
+  const receivedEmail = currentUserEmail;
+  const findUser = await UserRegister.findOne({ email: receivedEmail });
+  //console.log(findUser);
+  if (findUser?.role !== 'admin') {
+    throw new Error('You are not an admin for block A User!');
+  }
+
   const result = await UserRegister.findByIdAndUpdate(
     userId,
     { isBlocked: true },
@@ -10,12 +18,18 @@ const blockUserByAdminIntoDb = async (userId: string) => {
   return result;
 };
 
-const deleteBlogByAdminIntoDb = async (id: string) => {
+const deleteBlogByAdminIntoDB = async (id: string) => {
+  const receivedEmail = currentUserEmail;
+  const findUser = await UserRegister.findOne({ email: receivedEmail });
+  //console.log(findUser);
+  if (findUser?.role !== 'admin') {
+    throw new Error('You are not an admin !');
+  }
   const result = await BlogPost.findByIdAndDelete(id);
   return result;
 };
 
-export const adminService = {
-  deleteBlogByAdminIntoDb,
-  blockUserByAdminIntoDb,
+export const adminServices = {
+  deleteBlogByAdminIntoDB,
+  blockUserByAdminIntoDB,
 };
